refactor : Class name changed successfully - User -> Member
