+++ src/main/java/com/onulstore/domain/member/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.onulstore.domain.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+
+}
